Add  another version of func:generate_path()
For previous implementation, RRTTree::generate_path() generates the path from newest RRTNode from RRTTree to the beginning. So, when random node is added to one tree and some node of the other tree can be connected with added node coincidentally,  the node of the other tree is not newest and different node which is newest in the other tree is connected. I fixed it.